+++ Application/Request/PlatilloRequest.cs
@@ -3,6 +3,6 @@
     public class PlatilloRequest
     {
         public string descripcion { get; set; }
-        public double precio { get; set; }
+        public decimal precio { get; set; }
     }
 }
